Fixed all warnings and compilation issues that some systems had
Also fixed CPU exception handler, the error code was passed and not the interrupt number. In many cases, the exceptions don't even have error codes and the CPU pushes a 0x00 only to keep the stack aligned. This caused some exceptions to be reported falsely.

+++ src/Core/Interrupts/GDT.c
@@ -60,7 +60,7 @@ void InitGDT()
     GDTSetEntry(8, 0xFFFF, 0x00000000, 0xF3, 0xAF);
 
     GDTDesriptor.Size = sizeof(struct GDTEntry) * 9 - 1;
-    GDTDesriptor.Offset = &GDTEntries;
+    GDTDesriptor.Offset = (int64_t)&GDTEntries;
     asm volatile ("lgdt %0" :: "m"(GDTDesriptor) : "memory");
     GDTInitialized = true;
 }
